apps/authentications/serializers.py
from rest_framework import serializers

from apps.authentications.models import User


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username']
        read_only_fields = ['id']
