requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.10.0
gunicorn==20.1.0
Markdown==3.3.4
mysqlclient==2.0.3
pika==1.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
